AI_agent/requirements.txt
fastapi
uvicorn
langchain
chromadb
python-docx
python-dotenv
openai
apscheduler
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
slack-sdk
beautifulsoup4
pypdf
sentence-transformers
streamlit
pandas
numpy
python-multipart
